backend/requirements.txt
Flask>=2.0
Flask-SQLAlchemy
Flask-Cors
PyJWT
Werkzeug
